backend/apm_framework/requirements.txt
pymongo>=4.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
bson>=0.5.10
typing-extensions>=4.5.0
asyncio>=3.4.3
aiohttp>=3.8.4
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
logging>=0.4.9 